made a new ice block script
