+++ MeltdownUnity/Assets/Scenes/Testing Scenes/Soph.unity
@@ -952,17 +952,21 @@ PrefabInstance:
     serializedVersion: 3
     m_TransformParent: {fileID: 0}
     m_Modifications:
+    - target: {fileID: 4756320909481813501, guid: cdd3879758f497a4da8f09c0cdbe24d5, type: 3}
+      propertyPath: objectSetScale
+      value: 2
+      objectReference: {fileID: 0}
     - target: {fileID: 5870988160846128234, guid: cdd3879758f497a4da8f09c0cdbe24d5, type: 3}
       propertyPath: m_LocalPosition.x
       value: -0.13374066
       objectReference: {fileID: 0}
     - target: {fileID: 5870988160846128234, guid: cdd3879758f497a4da8f09c0cdbe24d5, type: 3}
       propertyPath: m_LocalPosition.y
-      value: -3.24
+      value: -3.46
       objectReference: {fileID: 0}
     - target: {fileID: 5870988160846128234, guid: cdd3879758f497a4da8f09c0cdbe24d5, type: 3}
       propertyPath: m_LocalPosition.z
-      value: 3.8050997
+      value: 2.22
       objectReference: {fileID: 0}
     - target: {fileID: 5870988160846128234, guid: cdd3879758f497a4da8f09c0cdbe24d5, type: 3}
       propertyPath: m_LocalRotation.w
